// FrontEnd/src/app/services/api-backend/model/asignacionesCama.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Caso } from './caso';
import { Cama } from './cama';


export interface AsignacionesCama { 
    idAsignacionCama?: number;
    idCaso?: number;
    idCama?: number;
    fechaInicio?: string;
    fechaFin?: string | null;
    estado?: string | null;
    idCamaNavigation?: Cama;
    idCasoNavigation?: Caso;
}