// FrontEnd/src/app/services/api-backend/model/cama.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Habitacione } from './habitacione';
import { AsignacionesCama } from './asignacionesCama';


export interface Cama { 
    idCama?: number;
    idHabitacion?: number;
    nombre?: string | null;
    idHabitacionNavigation?: Habitacione;
    asignacionesCamas?: Array<AsignacionesCama> | null;
}