// FrontEnd/src/app/services/api-backend/model/habitacione.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Sucursale } from './sucursale';
import { TiposHabitacion } from './tiposHabitacion';
import { Cama } from './cama';


export interface Habitacione { 
    idHabitacion?: number;
    idTipoHabitacion?: number;
    idSucursal?: number;
    nombre?: string | null;
    idSucursalNavigation?: Sucursale;
    idTipoHabitacionNavigation?: TiposHabitacion;
    camas?: Array<Cama> | null;
}