// FrontEnd/src/app/services/api-backend/model/sucursale.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Municipio } from './municipio';
import { Clinica } from './clinica';
import { Habitacione } from './habitacione';
import { Bodega } from './bodega';


export interface Sucursale { 
    idSucursal?: number;
    idMunicipio?: number;
    nombre?: string | null;
    direccion?: string | null;
    telefono?: string | null;
    estado?: string | null;
    idMunicipioNavigation?: Municipio;
    bodegas?: Array<Bodega> | null;
    clinicas?: Array<Clinica> | null;
    habitaciones?: Array<Habitacione> | null;
}